Create README.md
This article proposes a novel phase-field length scale insensitive model for fatigue failure in brittle materials. In the proposed model, we incorporate a necessary fatigue-related parameter to define the fatigue threshold energy as a function of the fracture strength and make the mechanical response of a material insensitive to the phase-field length scale. In the proposed model, we derive the governing partial differential equations by invoking the virtual power principle and assume constitutive relations for the thermodynamic fluxes on satisfying the thermodynamic laws. We provide a consistent derivation for determining the parameters that appear in the degradation function. We demonstrate the efficacy of the proposed model by generating phase-field length scale insensitive response in terms of crack length and maximum amplitude of load versus number of cycles for a few representative numerical examples, viz. a three-point bending test, a single-edge and a double-edge notched plate under low cycle fatigue. The numerical results highlight excellent insensitivity of the global mechanical response to the phase-field length scale parameter, validating the robustness of the proposed model. For numerical implementation, we have utilized an open-source finite element toolbox called Gridap, available in a high-performance programming language Julia, that facilitates third-party verification, promotes transparency and reproducibility, and sets a benchmark for efficient open-source code development in the scientific community.